Add files via upload
In the digital age, the rapid spread of misinformation poses a significant threat to public discourse, democracy, and societal well-being. The Fake News Detection project aims to combat the dissemination of fake news by leveraging machine learning techniques, specifically the TF-IDF vectorizer and the Passive-Aggressive Classifier. This project helps verify the legitimacy of information and mitigates the harmful effects of misinformation.